requirements.txt
fastapi
uvicorn[standard]
twilio
openai
deepgram-sdk
pybase64
numpy
elevenlabs
scipy
audioop-lts
groq
# Google Calendar dependencies
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client